Split dialog element builders
